all five flink examples passing
