+++ flink/flink-twitter-collection/src/test/resources/FlinkTwitterFollowingPipelineFollowersIT.conf
@@ -10,7 +10,11 @@ destination {
   path = "target/test-classes"
   writerPath = "FlinkTwitterFollowingPipelineFollowersIT"
 }
-twitter.endpoint = friends
+twitter {
+  endpoint = followers
+  ids_only = true
+  max_items = 5000
+}
 providerWaitMs = 1000
 local = true
 test = true
