+++ flink/flink-twitter-collection/src/test/resources/FlinkTwitterFollowingPipelineFriendsIT.conf
@@ -10,7 +10,10 @@ destination {
   path = "target/test-classes"
   writerPath = "FlinkTwitterFollowingPipelineFriendsIT"
 }
-twitter.endpoint = friends
+twitter {
+  endpoint = friends
+  ids_only = true
+}
 providerWaitMs = 1000
 local = true
 test = true
