+++ flink/flink-twitter-collection/src/test/resources/FlinkTwitterUserInformationPipelineIT.conf
@@ -2,7 +2,7 @@ source {
   fields = ["ID"]
   scheme = file
   path = "target/test-classes"
-  readerPath = "asf.txt"
+  readerPath = "1000twitterids.txt"
 }
 destination {
   fields = ["DOC"]
