+++ flink/flink-twitter-collection/src/test/scala/org/apache/streams/examples/flink/twitter/test/FlinkTwitterFollowingPipelineFriendsIT.scala
@@ -9,6 +9,7 @@ import org.apache.streams.config.{ComponentConfigurator, StreamsConfiguration, S
 import org.apache.streams.examples.flink.twitter.collection.FlinkTwitterUserInformationPipeline._
 import org.apache.streams.examples.flink.twitter.collection.{FlinkTwitterFollowingPipeline, FlinkTwitterSpritzerPipeline}
 import org.apache.streams.hdfs.{HdfsConfiguration, HdfsReaderConfiguration, HdfsWriterConfiguration}
+import org.junit.Ignore
 import org.slf4j.{Logger, LoggerFactory}
 import org.testng.annotations.Test
 
@@ -21,42 +22,14 @@ import org.scalatest.time.SpanSugar._
 /**
   * Created by sblackmon on 3/13/16.
   */
-class FlinkTwitterFollowingPipelineIT extends FlatSpec {
+class FlinkTwitterFollowingPipelineFriendsIT extends FlatSpec {
 
-  private val LOGGER: Logger = LoggerFactory.getLogger(classOf[FlinkTwitterFollowingPipelineIT])
+  private val LOGGER: Logger = LoggerFactory.getLogger(classOf[FlinkTwitterFollowingPipelineFriendsIT])
 
   import FlinkTwitterFollowingPipeline._
 
-  @Test(enabled = false)
+  @Test
   def flinkTwitterFollowersPipelineFriendsIT = {
-
-    val reference: Config = ConfigFactory.load()
-    val conf_file: File = new File("target/test-classes/FlinkTwitterFollowingPipelineFollowersIT.conf")
-    assert(conf_file.exists())
-    val testResourceConfig: Config = ConfigFactory.parseFileAnySyntax(conf_file, ConfigParseOptions.defaults().setAllowMissing(false));
-
-    val typesafe: Config = testResourceConfig.withFallback(reference).resolve()
-    val streams: StreamsConfiguration = StreamsConfigurator.detectConfiguration(typesafe)
-    val testConfig = new ComponentConfigurator(classOf[TwitterFollowingPipelineConfiguration]).detectConfiguration(typesafe)
-
-    setup(testConfig)
-
-    val job = new FlinkTwitterFollowingPipeline(config = testConfig)
-    val jobThread = new Thread(job)
-    jobThread.start
-    jobThread.join
-
-    eventually (timeout(60 seconds), interval(1 seconds)) {
-      assert(Files.exists(Paths.get(testConfig.getDestination.getPath + "/" + testConfig.getDestination.getWriterPath)))
-      assert(
-        Source.fromFile(testConfig.getDestination.getPath + "/" + testConfig.getDestination.getWriterPath, "UTF-8").getLines.size
-          > 90)
-    }
-
-  }
-
-  @Test(enabled = false)
-  def flinkTwitterFollowersPipelineFollowersIT = {
 
     val reference: Config = ConfigFactory.load()
     val conf_file: File = new File("target/test-classes/FlinkTwitterFollowingPipelineFriendsIT.conf")
@@ -78,7 +51,7 @@ class FlinkTwitterFollowingPipelineIT extends FlatSpec {
       assert(Files.exists(Paths.get(testConfig.getDestination.getPath + "/" + testConfig.getDestination.getWriterPath)))
       assert(
         Source.fromFile(testConfig.getDestination.getPath + "/" + testConfig.getDestination.getWriterPath, "UTF-8").getLines.size
-          > 500)
+          > 90)
     }
 
   }
